backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
alembic
celery
redis
pydantic-settings
python-dotenv
anthropic
openai
httpx
python-telegram-bot==20.8
passlib[bcrypt]
python-jose[cryptography]
pyotp
qrcode
email-validator
pytz
fitparse
pandas
python-multipart